use alert-dialog

+++ apps/example-think-python/src/components/note/node-delete.tsx
@@ -1,18 +1,20 @@
 "use client";
 import { TrashIcon } from "lucide-react";
-import {
-	Dialog,
-	DialogTrigger,
-	DialogContent,
-	DialogHeader,
-	DialogTitle,
-	DialogFooter,
-} from "@/components/ui/dialog";
 
-import { Button } from "../client-components";
+import {
+	Button,
+	AlertDialog,
+	AlertDialogTrigger,
+	AlertDialogContent,
+	AlertDialogHeader,
+	AlertDialogTitle,
+	AlertDialogDescription,
+	AlertDialogFooter,
+	AlertDialogAction,
+	AlertDialogCancel,
+} from "../client-components";
 import { Spinner } from "../spinner";
 import { useState } from "react";
-import { flushSync } from "react-dom";
 
 type Props = {
 	// need this prop to tell note-card the modal is open
@@ -26,37 +28,41 @@ export const NoteDelete = ({ open, onOpenChange, onDelete }: Props) => {
 	const [isLoading, setIsLoading] = useState(false);
 
 	return (
-		<Dialog
+		<AlertDialog
 			open={open}
 			onOpenChange={(val) => {
 				onOpenChange(val);
 			}}
 		>
-			<DialogTrigger asChild>
+			<AlertDialogTrigger>
 				<Button variant="ghost" size="sm" type="button">
 					<TrashIcon className="w-4 h-4" />
 				</Button>
-			</DialogTrigger>
-			<DialogContent>
-				<DialogHeader>
-					<DialogTitle>Delete this Note?</DialogTitle>
-				</DialogHeader>
-				<DialogFooter>
-					<Button variant="outline" onClick={() => onOpenChange(false)}>
-						Cancel
-					</Button>
-					<Button
+			</AlertDialogTrigger>
+			<AlertDialogContent>
+				<AlertDialogHeader>
+					<AlertDialogTitle>
+						Are you sure you want to delete this note?
+					</AlertDialogTitle>
+					<AlertDialogDescription>
+						This action cannot be undone.
+					</AlertDialogDescription>
+				</AlertDialogHeader>
+				<AlertDialogFooter>
+					<AlertDialogCancel>Cancel</AlertDialogCancel>
+					<AlertDialogAction
 						disabled={isLoading}
 						onClick={async () => {
 							setIsLoading(true);
 							await onDelete();
 							setIsLoading(false);
 						}}
+						className="bg-red-600 focus:ring-red-600"
 					>
 						{isLoading ? <Spinner /> : <span>Delete</span>}
-					</Button>
-				</DialogFooter>
-			</DialogContent>
-		</Dialog>
+					</AlertDialogAction>
+				</AlertDialogFooter>
+			</AlertDialogContent>
+		</AlertDialog>
 	);
 };
